move squin to stim rewrite to stim dialect, split out wrap analysis infrastructure
