--- src/bloqade/squin/passes/__init__.py
@@ -1,1 +1,0 @@
-from .stim import SquinToStim as SquinToStim
