+++ src/bloqade/squin/rewrite/__init__.py
@@ -1,11 +1,6 @@
-from .wire_to_stim import SquinWireToStim as SquinWireToStim
-from .qubit_to_stim import SquinQubitToStim as SquinQubitToStim
-from .squin_measure import SquinMeasureToStim as SquinMeasureToStim
 from .wrap_analysis import (
     SitesAttribute as SitesAttribute,
     AddressAttribute as AddressAttribute,
-    WrapSquinAnalysis as WrapSquinAnalysis,
-)
-from .wire_identity_elimination import (
-    SquinWireIdentityElimination as SquinWireIdentityElimination,
+    WrapOpSiteAnalysis as WrapOpSiteAnalysis,
+    WrapAddressAnalysis as WrapAddressAnalysis,
 )
