+++ src/bloqade/stim/__init__.py
@@ -1,4 +1,4 @@
-from . import emit as emit, parse as parse, dialects as dialects
+from . import emit as emit, parse as parse, passes as passes, dialects as dialects
 from .groups import main as main
 from ._wrappers import (
     h as h,
