+++ src/bloqade/stim/passes/__init__.py
@@ -1,0 +1,1 @@
+from .squin_to_stim import SquinToStim as SquinToStim
