+++ src/bloqade/stim/passes/squin_to_stim.py
@@ -12,17 +12,14 @@ from kirin.ir.method import Method
 from kirin.passes.abc import Pass
 from kirin.rewrite.abc import RewriteResult
 
-from bloqade.squin.rewrite import (
+from bloqade.stim.rewrite import (
     SquinWireToStim,
     SquinQubitToStim,
-    WrapSquinAnalysis,
     SquinMeasureToStim,
     SquinWireIdentityElimination,
 )
+from bloqade.squin.rewrite import WrapAddressAnalysis
 from bloqade.analysis.address import AddressAnalysis
-from bloqade.squin.analysis.nsites import (
-    NSitesAnalysis,
-)
 
 
 @dataclass
@@ -36,17 +33,12 @@ class SquinToStim(Pass):
         # Get necessary analysis results to plug into hints
         address_analysis = AddressAnalysis(mt.dialects)
         address_frame, _ = address_analysis.run_analysis(mt)
-        site_analysis = NSitesAnalysis(mt.dialects)
-        sites_frame, _ = site_analysis.run_analysis(mt)
 
         # Wrap Rewrite + SquinToStim can happen w/ standard walk
         rewrite_result = (
             Walk(
                 Chain(
-                    WrapSquinAnalysis(
-                        address_analysis=address_frame.entries,
-                        op_site_analysis=sites_frame.entries,
-                    ),
+                    WrapAddressAnalysis(address_frame.entries),
                     SquinQubitToStim(),
                     SquinWireToStim(),
                     SquinMeasureToStim(),  # reduce duplicated logic, can split out even more rules later
