+++ src/bloqade/stim/rewrite/qubit_to_stim.py
@@ -2,8 +2,8 @@ from kirin import ir
 from kirin.rewrite.abc import RewriteRule, RewriteResult
 
 from bloqade.squin import op, qubit
-from bloqade.squin.rewrite.wrap_analysis import AddressAttribute
-from bloqade.squin.rewrite.stim_rewrite_util import (
+from bloqade.squin.rewrite import AddressAttribute
+from bloqade.stim.rewrite.util import (
     SQUIN_STIM_GATE_MAPPING,
     rewrite_Control,
     insert_qubit_idx_from_address,
