+++ src/bloqade/stim/rewrite/wire_identity_elimination.py
